100 - move data holders to own folder
