+++ tests/tel2puml/otel_to_pv/test_find_unique_graphs.py
@@ -19,7 +19,7 @@ from tel2puml.otel_to_pv.find_unique_graphs import (
     get_unique_graph_job_ids_per_job_name,
     find_unique_graphs
 )
-from tel2puml.otel_to_pv.otel_ingestion.otel_data_holder import (
+from tel2puml.otel_to_pv.data_holders.otel_data_holder import (
     DataHolder, SQLDataHolder
 )
 from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
