+++ tests/tel2puml/otel_to_pv/test_ingest_data.py
@@ -16,7 +16,7 @@ from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
     NodeModel,
     IngestTypes
 )
-from tel2puml.otel_to_pv.otel_ingestion.otel_data_holder import (
+from tel2puml.otel_to_pv.data_holders.otel_data_holder import (
     SQLDataHolder,
 )
 from tel2puml.otel_to_pv.otel_ingestion.otel_data_source import (
