+++ tests/tel2puml/otel_to_pv/test_otel_data_holder.py
@@ -17,7 +17,7 @@ from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
     Base,
     NODE_ASSOCIATION,
 )
-from tel2puml.otel_to_pv.otel_ingestion.otel_data_holder import (
+from tel2puml.otel_to_pv.data_holders.otel_data_holder import (
     SQLDataHolder,
 )
